pyproject.toml
[project]
name = "mag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "asana>=5.2.2",
    "beautifulsoup4>=4.14.2",
    "chromadb>=1.3.4",
    "flask>=3.1.2",
    "langchain>=1.0.7",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.3",
    "langchain-text-splitters>=1.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
